Replace `dtslint` with `tsd`
