+++ .github/workflows/node.js.yml
@@ -33,8 +33,8 @@ jobs:
       - name: Run tests
         run: npm test
 
-  dtslint:
-    name: Validate type definitions
+  check-types:
+    name: Check type definitions
     runs-on: ubuntu-latest
 
     steps:
@@ -50,5 +50,5 @@ jobs:
       - name: Install dependencies
         run: npm ci
 
-      - name: Validate type definitions
-        run: npm run dtslint
+      - name: Check type definitions
+        run: npm run check-types
