--- tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "lib": ["es2015"],
-    "moduleResolution": "node",
-    "strict": true,
-    "noEmit": true,
-    "baseUrl": ".",
-    "paths": {
-      "classnames": ["."],
-      "classnames/*": ["./*"]
-    }
-  }
-}
