--- tslint.json
@@ -1,3 +1,0 @@
-{
-  "extends": "dtslint/dtslint.json"
-}
